Move ChunkedVector to its own header
